Move rest/domain/Stream into src/stream.
