+++ src/rest/StreamEndpoints.js
@@ -4,10 +4,10 @@ import { Agent as HttpsAgent } from 'https'
 import qs from 'qs'
 import debugFactory from 'debug'
 
-import { validateOptions } from '../stream/utils'
 import { getEndpointUrl } from '../utils'
+import { validateOptions } from '../stream/utils'
+import Stream from '../stream'
 
-import Stream from './domain/Stream'
 import authFetch from './authFetch'
 
 const debug = debugFactory('StreamrClient')
